--- packages/runtime/fixtures/configs/monorepo-client-without-id.json
@@ -1,23 +1,0 @@
-{
-  "$schema": "https://platformatic.dev/schemas/v0.20.0/runtime",
-  "entrypoint": "serviceApp",
-  "hotReload": false,
-  "autoload": {
-    "path": "../monorepo",
-    "exclude": ["docs", "composerApp"],
-    "mappings": {
-      "serviceApp": {
-        "id": "serviceApp",
-        "config": "platformatic.service-client-without-id.json"
-      },
-      "serviceAppWithLogger": {
-        "id": "with-logger",
-        "config": "platformatic.service.json"
-      },
-      "serviceAppWithMultiplePlugins": {
-        "id": "multi-plugin-service",
-        "config": "platformatic.service.json"
-      }
-    }
-  }
-}
