--- packages/runtime/fixtures/monorepo/serviceApp/platformatic.service-client-without-id.json
@@ -1,21 +1,0 @@
-{
-  "$schema": "https://platformatic.dev/schemas/v0.20.0/service",
-  "server": {
-    "hostname": "127.0.0.1",
-    "port": 0
-  },
-  "service": {
-    "openapi": true
-  },
-  "plugins": {
-    "paths": [
-      "plugin.js"
-    ]
-  },
-  "clients": [
-    {
-      "path": "./with-logger",
-      "url": "{PLT_WITH_LOGGER_URL}"
-    }
-  ]
-}
